refactor: 도메인, 컨트롤러, 리포지토리 코드 정리
