+++ src/main/java/nextstep/reservations/repository/reservation/JdbcTemplateReservationRepository.java
@@ -1,11 +1,10 @@
-package nextstep.reservations.domain.repository.reservation;
+package nextstep.reservations.repository.reservation;
 
 import nextstep.reservations.domain.entity.reservation.Reservation;
 import nextstep.reservations.domain.entity.theme.Theme;
 import nextstep.reservations.exceptions.reservation.exception.DuplicateReservationException;
 import nextstep.reservations.exceptions.reservation.exception.NoSuchReservationException;
 import nextstep.reservations.exceptions.theme.exception.NoSuchThemeException;
-import org.springframework.context.annotation.Primary;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -18,7 +17,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Objects;
 
 @Repository
-@Primary
+//@Primary
 public class JdbcTemplateReservationRepository implements ReservationRepository{
     private static final String INSERT_ONE_QUERY = "INSERT INTO reservation (date, time, name, theme_name, theme_desc, theme_price) VALUES (?, ?, ?, ?, ?, ?)";
     private static final String FIND_BY_ID_QUERY = "SELECT * FROM reservation WHERE id = ?";
