+++ src/test/java/nextstep/reservations/ReservationTest.java
@@ -1,11 +1,10 @@
 package nextstep.reservations;
 
 import io.restassured.RestAssured;
-import nextstep.reservations.domain.repository.reservation.ReservationRepository;
+import nextstep.reservations.repository.reservation.ReservationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
